refactor(frontend): refactor graphql query into hooks
