+++ frontend/src/pages/InstancesPage.tsx
@@ -1,37 +1,23 @@
-import { gql, useQuery } from "@apollo/client";
 import { Table } from "../components/Table";
-import { queryPollIntervalMs } from "../utils/constants";
+import { useQueryInstances } from "../hooks/useQueryInstances";
 
 export default function InstancesPage() {
-  const INSTANCES = gql`
-    query Instances {
-      instances {
-        instanceKey
-        process {
-          bpmnProcessId
-        }
-        status
-        startTime
-      }
-    }
-  `;
-  const { data } = useQuery(INSTANCES, {
-    pollInterval: queryPollIntervalMs,
-  });
+  const { instances } = useQueryInstances();
+
   return (
     <Table
       orientation="horizontal"
       header={["Instance Key", "BPMN Process ID", "Status", "Start Time"]}
       content={
-        data
-          ? data.instances.map(
+        instances
+          ? instances.map(
               ({
                 instanceKey,
                 process: { bpmnProcessId },
                 status,
                 startTime,
               }: {
-                instanceKey: string;
+                instanceKey: number;
                 process: { bpmnProcessId: string };
                 status: string;
                 startTime: string;
